requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2018.10.15
cffi==1.16.0
chardet==3.0.4
defusedxml==0.5.0
Django==3.2.7
django-cors-headers==4.3.1
django-filter==2.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
idna==2.7
inflection==0.5.1
mysqlclient==2.2.0
loguru==0.5.3
oauthlib==2.1.0
packaging==23.2
py-bcrypt==0.4
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==1.24.2